use artifactory repo instead of harbor to follow srp rule

+++ hack/verify-helm.sh
@@ -18,7 +18,7 @@ THIS_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 _BINDIR="$THIS_DIR/.bin"
 # Must be an exact match, as the generated YAMLs may not be consistent across
 # versions
-_HELM_VERSION="v3.8.1"
+_HELM_VERSION="v3.12.3"
 
 # Ensure the helm tool exists and is the correct version, or install it
 verify_helm() {
@@ -31,6 +31,7 @@ verify_helm() {
         # Should work with:
         #  - v3.8.1
         #  - v3.8.1+g5cb9af4
+        helm_version="${helm_version%+*}"
         helm_version="${helm_version%+*}"
         if [ "${helm_version}" == "${_HELM_VERSION}" ]; then
             # If version is exact match, stop here.
@@ -62,11 +63,16 @@ verify_helm() {
     esac
     
     >&2 echo "Installing helm"
-    local helm_url="https://get.helm.sh/helm-${_HELM_VERSION}-${ostype}-${arch}.tar.gz"
-    curl -sLo helm.tar.gz "${helm_url}" || return 1
     mkdir -p "$_BINDIR" || return 1
-    tar -xzf helm.tar.gz -C "$_BINDIR" --strip-components=1 "${ostype}-${arch}/helm" || return 1
-    rm -f helm.tar.gz
+    if [ -z $1 ]; then
+        local helm_url="https://get.helm.sh/helm-${_HELM_VERSION}-${ostype}-${arch}.tar.gz"
+        curl -sLo helm.tar.gz "${helm_url}" || return 1
+        tar -xzf helm.tar.gz -C "$_BINDIR" --strip-components=1 "${ostype}-${arch}/helm" || return 1
+        rm -f helm.tar.gz
+    else
+        cp $1 "$_BINDIR"
+        chmod +x "$_BINDIR"/helm
+    fi
     helm="$_BINDIR/helm"
     echo "$helm"
     return 0
